+++ frontend/app/page.tsx
@@ -1,8 +1,5 @@
 "use client";
 
-import { useState, useEffect, useCallback } from "react";
-import { Restaurant, SearchResults } from "./types/restaurant";
-import { useRouter, useSearchParams } from "next/navigation";
 import LocationFinder from "./components/restaurants/LocationFinder";
 import SearchForm from "./components/restaurants/SearchForm";
 import LoadingIndicator from "./components/ui/LoadingIndicator";
@@ -10,16 +7,17 @@ import RestaurantCard from "./components/restaurants/RestaurantCard";
 import Pagination from "./components/restaurants/Pagination";
 import ErrorMessage from "./components/ui/ErrorMessage";
 import { useGeolocation } from "./hooks/useGeolocation";
+import { useRestaurantSearch } from "./hooks/useRestaurantSearch";
+import { useMemo } from "react";
 
 export default function SearchPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
+  // このコンポーネントは、ユーザーの位置情報をもとにレストランを検索し結果を表示する
+  // 主な機能:
+  // 1. 現在地の取得と表示
+  // 2. 検索半径の指定と検索実行
+  // 3. 検索結果の表示とページネーション
 
-  // URLからページ番号を取得（ない場合は1をデフォルト値とする）
-  const pageFromUrl = searchParams.get("page")
-    ? parseInt(searchParams.get("page") as string)
-    : 1;
-
+  // 位置情報取得のカスタムフック - ユーザーの緯度・経度を提供
   const {
     location,
     isLoading: isLocationLoading,
@@ -27,126 +25,22 @@ export default function SearchPage() {
     getCurrentLocation,
   } = useGeolocation();
 
-  const [radius, setRadius] = useState<string>(
-    searchParams.get("radius") || "3"
-  );
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [error, setError] = useState<string | null>(null);
-  const [restaurants, setRestaurants] = useState<Restaurant[]>([]);
-  const [resultsInfo, setResultsInfo] = useState<{
-    available: number;
-    returned: number;
-    start: number;
-  } | null>(null);
-  const perPage = 20; // 1ページあたりの表示件数
+    // レストラン検索のカスタムフック - 位置情報をもとにAPIからデータを取得
+  const {
+    pageFromUrl,     // URLから取得したページ番号
+    radius,          // 検索半径
+    isLoading,       // データ取得中のローディング状態
+    error,           // 検索エラー
+    restaurants,     // 取得したレストラン一覧
+    resultsInfo,     // 検索結果のメタデータ（総件数、表示件数など）
+    totalPages,      // 総ページ数
+    searchRestaurants, // 検索を実行する関数
+    handleSubmit,     // 検索フォーム送信ハンドラ
+    handleRadiusChange, // 検索半径変更ハンドラ
+    getPageNumbers,   // ページネーション用の表示ページ番号配列を生成
+  } = useRestaurantSearch({ location });
 
-  // クエリパラメータを更新する関数（位置情報は含めない）
-  const updateQueryParams = useCallback((page: number, newRadius?: string) => {
-    // 現在のURLパラメータを取得
-    const params = new URLSearchParams(searchParams.toString());
-
-    // パラメータ更新
-    params.set("page", page.toString());
-    if (newRadius) params.set("radius", newRadius);
-
-    // URLを更新（ページ遷移せずにURLのみ変更）
-    router.push(`?${params.toString()}`, { scroll: false });
-  }, [searchParams, router]);
-
-  // 検索を実行
-  const searchRestaurants = useCallback(async (page: number = pageFromUrl) => {
-    if (!location.lat || !location.lng) {
-      setError("位置情報を先に取得してください");
-      return;
-    }
-
-    setIsLoading(true);
-    setError(null);
-
-    // URLのページ番号を更新
-    updateQueryParams(page, radius);
-
-    // 開始位置を計算
-    const start = (page - 1) * perPage + 1;
-
-    try {
-      const response = await fetch(
-        `/api/restaurants/search?lat=${location.lat}&lng=${location.lng}&radius=${radius}&start=${start}&count=${perPage}`
-      );
-
-      if (!response.ok) {
-        throw new Error("レストラン検索に失敗しました");
-      }
-
-      const data: SearchResults = await response.json();
-      setRestaurants(data.results.shop || []);
-      setResultsInfo({
-        available: data.results.results_available,
-        returned: parseInt(data.results.results_returned),
-        start: data.results.results_start,
-      });
-    } catch (error) {
-      console.error("検索エラー:", error);
-      setError("レストランの検索中にエラーが発生しました");
-    } finally {
-      setIsLoading(false);
-    }
-  }, [location.lat, location.lng, radius, perPage, updateQueryParams, pageFromUrl]);
-
-  // 検索フォームの送信ハンドラ
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    searchRestaurants(1); // 検索時は1ページ目から表示
-  };
-
-  // 検索範囲変更時のハンドラ
-  const handleRadiusChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    const newRadius = e.target.value;
-    setRadius(newRadius);
-  };
-
-  const totalPages = resultsInfo
-    ? Math.ceil(resultsInfo.available / perPage)
-    : 0;
-
-  // ページネーションのコントロールに使用する配列の生成
-  const getPageNumbers = useCallback(() => {
-    // 表示するページネーション数
-    const maxVisiblePages = 5;
-    let startPage: number;
-    let endPage: number;
-
-    if (totalPages <= maxVisiblePages) {
-      startPage = 1;
-      endPage = totalPages;
-    } else {
-      const middleOffset = Math.floor(maxVisiblePages / 2);
-
-      if (pageFromUrl <= middleOffset + 1) {
-        startPage = 1;
-        endPage = maxVisiblePages;
-      } else if (pageFromUrl >= totalPages - middleOffset) {
-        startPage = totalPages - maxVisiblePages + 1;
-        endPage = totalPages;
-      } else {
-        startPage = pageFromUrl - middleOffset;
-        endPage = pageFromUrl + middleOffset;
-      }
-    }
-
-    return Array.from(
-      { length: endPage - startPage + 1 },
-      (_, i) => startPage + i
-    );
-  }, [totalPages, pageFromUrl]);
-
-  // コンポーネントがマウントされたときとURLのページ番号が変わったときに検索を実行
-  useEffect(() => {
-    // 位置情報がある場合のみ検索を実行
-    if (location.lat && location.lng) {
-      searchRestaurants(pageFromUrl);
-    }
-  }, [pageFromUrl, location.lat, location.lng, searchRestaurants]); // URLのページが変わったら、または位置情報が設定されたら再検索
+  const pageNumbers = useMemo(() => getPageNumbers(), [getPageNumbers]);
 
   return (
     <div className="bg-gray-50 min-h-screen w-full">
@@ -157,7 +51,7 @@ export default function SearchPage() {
 
         {/* 現在地取得と検索フォームを横並びに */}
         <div className="grid grid-cols-1 lg:grid-cols-2 gap-4 mb-6">
-          {/* 現在地取得エリア */}
+          {/* 現在地取得UI - 位置情報の表示と取得ボタン */}
           <LocationFinder
             location={location}
             isLoading={isLocationLoading}
@@ -174,13 +68,15 @@ export default function SearchPage() {
           />
         </div>
 
-        {/* エラー表示 - グローバルに */}
-        {( error || locationError ) && (<ErrorMessage message={error || locationError || ""} />)}
+        {/* エラー表示 - 位置情報または検索でエラーが発生した場合 */}
+        {(error || locationError) && (
+          <ErrorMessage message={(error || locationError || "")} />
+        )}
 
-        {/* ローディングインジケータ - 非モーダル */}
+        {/* エラー表示 - 位置情報または検索でエラーが発生した場合 */}
         {isLoading && <LoadingIndicator />}
 
-        {/* 検索結果 */}
+        {/* 検索結果のメタ情報表示 - 検索完了後のみ表示 */}
         {resultsInfo && !isLoading && (
           <div className="mb-6 bg-white p-4 rounded-lg shadow-sm border border-gray-100">
             <div className="flex items-center justify-between">
@@ -197,29 +93,36 @@ export default function SearchPage() {
           </div>
         )}
 
-        {restaurants.length > 0 && !isLoading ? (
-          <>
-            {/* カードグリッド */}
-            <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 mb-8">
-              {restaurants.map((restaurant) => (
-                <RestaurantCard key={restaurant.id} restaurant={restaurant} />
-              ))}
-            </div>
+        {/* 条件分岐による結果表示 */}
+        {!isLoading && resultsInfo && (
+          restaurants.length > 0 ? (
+            <>
+              {/* レストランカードのグリッド表示 */}
+              <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 mb-8">
+                {restaurants.map((restaurant) => (
+                  <RestaurantCard 
+                    key={restaurant.id} 
+                    restaurant={restaurant} 
+                  />
+                ))}
+              </div>
 
-            {/* ページネーション */}
-            <Pagination
-              currentPage={pageFromUrl}
-              totalPages={totalPages}
-              isLoading={isLoading}
-              pageNumbers={getPageNumbers()}
-              onPageChange={searchRestaurants}
-            />
-          </>
-        ) : resultsInfo && !isLoading ? (
-          <div className="text-center py-10 bg-white rounded-lg shadow-sm border border-gray-100">
-            <p className="text-gray-600">検索結果がありません</p>
-          </div>
-        ) : null}
+              {/* ページネーションコントロール */}
+              <Pagination
+                currentPage={pageFromUrl}
+                totalPages={totalPages}
+                isLoading={isLoading}
+                pageNumbers={pageNumbers}
+                onPageChange={searchRestaurants}
+              />
+            </>
+          ) : (
+            // 検索結果が0件の場合のメッセージ
+            <div className="text-center py-10 bg-white rounded-lg shadow-sm border border-gray-100">
+              <p className="text-gray-600">検索結果がありません</p>
+            </div>
+          )
+        )}
       </main>
     </div>
   );
